pyproject.toml

[project]
name = "canonical-multipass-mcp-server"
version = "0.1.5"
description = "A MCP server for managing Canonical Multipass VMs."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Xiaodong Ye", email = "yeahdongcn@gmail.com" },
]
maintainers = [
    { name = "Xiaodong Ye", email = "yeahdongcn@gmail.com" },
]
keywords = ["ai", "agents", "mcp", "model-context-protocol", "canonical", "multipass", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.3",
    "httpx>=0.28.1",
    "grpcio-tools>=1.73.1",
]

[project.urls]
Homepage = "https://github.com/yeahdongcn/canonical-multipass-mcp-server"
Documentation = "https://github.com/yeahdongcn/canonical-multipass-mcp-server#readme"
Repository = "https://github.com/yeahdongcn/canonical-multipass-mcp-server.git"
Issues = "https://github.com/yeahdongcn/canonical-multipass-mcp-server/issues"
Changelog = "https://github.com/yeahdongcn/canonical-multipass-mcp-server/releases"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0"
]

[project.scripts]
canonical-multipass-mcp-server = "canonical_multipass_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["canonical_multipass_mcp"]

[tool.black]
line-length = 79
target-version = ['py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
